+++ src-core/src/globals.rs
@@ -1,4 +1,3 @@
-use config::Config;
 use std::sync::LazyLock;
 use tauri::AppHandle;
 use tauri_plugin_cli::Matches;
@@ -14,13 +13,3 @@ pub const APTABASE_HOST: &str = "https://aptabase.raphii.co";
 
 pub static TAURI_APP_HANDLE: LazyLock<Mutex<Option<AppHandle>>> = LazyLock::new(Default::default);
 pub static TAURI_CLI_MATCHES: LazyLock<Mutex<Option<Matches>>> = LazyLock::new(Default::default);
-pub static FLAGS: LazyLock<Mutex<Option<Config>>> = LazyLock::new(Default::default);
-
-#[allow(dead_code)]
-pub async fn is_flag_set(flag: &str) -> bool {
-    let flags = FLAGS.lock().await;
-    if let Some(flags) = flags.as_ref() {
-        return flags.get_bool(flag).unwrap_or(false);
-    }
-    false
-}