+++ src-core/src/openvr/commands.rs
@@ -11,7 +11,6 @@ use ovr_overlay as ovr;
 use substring::Substring;
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_set_app_framelimit(
     app_id: u32,
     limits: Option<OVRFrameLimits>,
@@ -20,25 +19,21 @@ pub async fn openvr_set_app_framelimit(
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_get_app_framelimit(app_id: u32) -> Result<Option<OVRFrameLimits>, String> {
     super::framelimiter::get_app_framelimits(app_id).await
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_set_init_delay_fix(enabled: bool) {
     *super::OVR_INIT_DELAY_FIX.lock().await = enabled;
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_get_devices() -> Vec<OVRDevice> {
     super::devices::get_devices().await
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_status() -> String {
     let status = super::OVR_STATUS.lock().await;
     let status_str = serde_json::to_string(&*status).unwrap();
@@ -46,43 +41,36 @@ pub async fn openvr_status() -> String {
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_set_analog_gain(analog_gain: f32) -> Result<(), String> {
     super::brightness_analog::set_analog_gain(analog_gain).await
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_get_analog_gain() -> Result<f32, String> {
     super::brightness_analog::get_analog_gain().await
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_set_supersample_scale(supersample_scale: Option<f32>) -> Result<(), String> {
     super::supersampling::set_supersample_scale(supersample_scale).await
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_get_supersample_scale() -> Result<Option<f32>, String> {
     super::supersampling::get_supersample_scale().await
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_set_fade_distance(fade_distance: f32) -> Result<(), String> {
     super::chaperone::set_fade_distance(fade_distance).await
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_get_fade_distance() -> Result<f32, String> {
     super::chaperone::get_fade_distance().await
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_set_analog_color_temp(
     temperature: Option<u32>,
 ) -> Result<(f64, f64, f64), String> {
@@ -90,7 +78,6 @@ pub async fn openvr_set_analog_color_temp(
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_set_image_brightness(
     brightness: f64,
     perceived_brightness_adjustment_gamma: Option<f64>,
@@ -100,7 +87,6 @@ pub async fn openvr_set_image_brightness(
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_launch_binding_configuration(show_on_desktop: bool) {
     let context = OVR_CONTEXT.lock().await;
     let mut input = match context.as_ref() {
@@ -120,7 +106,6 @@ pub async fn openvr_launch_binding_configuration(show_on_desktop: bool) {
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_is_dashboard_visible() -> bool {
     let context = OVR_CONTEXT.lock().await;
     let mut manager = match context.as_ref() {
@@ -131,7 +116,6 @@ pub async fn openvr_is_dashboard_visible() -> bool {
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_reregister_manifest() -> Result<(), String> {
     let ctx = OVR_CONTEXT.lock().await;
     let mut applications = ctx.as_ref().unwrap().applications_mngr();
@@ -182,7 +166,6 @@ pub async fn openvr_reregister_manifest() -> Result<(), String> {
 }
 
 #[tauri::command]
-#[oyasumivr_macros::command_profiling]
 pub async fn openvr_get_binding_origins(
     action_set_key: String,
     action_key: String,